move calculateAge to utils
